packages/guardrailgen-py/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "guardrailgen"
version = "0.0.1"
description = "OpenClaw-first guardrails test generator and proof artifacts"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
guardrailgen = "guardrailgen.cli:main"
